Add numpy/log-probability based HMM implementation
I removed some functions I didn't think we'd use (viterbi and
supervised learning). Every function should work as described in the
docstring (in particular, unsupervised_HMM should work the same as
before). Unsupervised learning results were confirmed to match the
checkpoints given for question 2D in set 6.

A, O, alphas, and betas have been converted to log-probabilities instead
of raw probabilities, to prevent underflow. This means that what used to
be self.A is now np.exp(self.A), etc.